sort out clojure pushTests
